+++ Serilog.Sinks.Console.Themes.UnitTests/GlobalUsings.cs
@@ -1,11 +1,16 @@
-﻿// Global using directives
+// Global using directives
 
 global using System;
 global using System.Collections.Generic;
 global using System.Drawing;
+global using System.IO;
 global using System.Linq;
 global using System.Reflection;
 global using FluentAssertions;
 global using NUnit.Framework;
+global using Serilog.Events;
+global using Serilog.Parsing;
 global using Serilog.Sinks.Console.Themes;
 global using Serilog.Sinks.SystemConsole.Themes;
+global using Serilog.Templates;
+global using Serilog.Templates.Themes;